requirements.txt
pandas
numpy
scikit-learn
flask
setuptools
google-cloud-storage
pyyaml